Replace invite command with link command in group module
